@@ -5,13 +5,14 @@
 	<artifactId>dev-ref</artifactId>
 	<version>5.12.2-Eval</version>
 	<properties>
-		<keikai.version>${project.version}</keikai.version>
+		<keikai.version>5.12.2-jakarta-FL-20230808</keikai.version>
+		<zk.jakarta.version>9.6.3-jakarta</zk.jakarta.version>
 		<zk.version>9.6.3-Eval</zk.version>
 		<jsf-api.version>2.2.20</jsf-api.version>
 		<maven.build.timestamp.format>yyyy-MM-dd</maven.build.timestamp.format>
 		<packname>-${project.version}-FL-${maven.build.timestamp}</packname>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>1.9</maven.compiler.source>
+		<maven.compiler.target>1.9</maven.compiler.target>
 	</properties>
 	<packaging>war</packaging>
 	<name>The Keikai Developer Reference Example Project</name>
